refactor: 기존 assembleReconstructBlockTree 를 domain.service로 옴겨 관심사 분리 및 책임을 분리하여 도메인 로직의 응집도 높임
