+++ src/main/java/com/joojoo/api/block/application/detail/BlockDtoAssemblerImpl.java
@@ -1,8 +1,6 @@
 package com.joojoo.api.block.application.detail;
 
 import com.joojoo.api.block.domain.model.entity.Block;
-import com.joojoo.api.block.presentation.dto.response.blockDetail.BlockResponse;
-import com.joojoo.api.block.presentation.dto.response.blockDetail.BlockResponseDto;
 import com.joojoo.api.block.presentation.dto.response.detail.BlockDetailResponseDto;
 import com.joojoo.api.blockTag.domain.model.entity.BlockTag;
 import com.joojoo.api.blockTicker.domain.model.entity.BlockTicker;
@@ -56,49 +54,6 @@ public class BlockDtoAssemblerImpl implements BlockDtoAssembler {
                         childCounts.getOrDefault(block.getId(), 0L)
                 ))
                 .toList();
-    }
-
-    @Override
-    public BlockResponse assembleReconstructBlockTree(List<Block> allBlocks) {
-        Map<Long, BlockResponseDto> dtoMap = getBlockResponseDtoMap(allBlocks);
-        connectNodes(allBlocks, dtoMap);
-        return BlockResponse.of(getRootBlocks(allBlocks, dtoMap));
-    }
-
-    @Override
-    public List<Long> toIds(List<Block> blockIds) {
-        return blockIds.stream()
-                .map(Block::getId)
-                .collect(Collectors.toList());
-    }
-
-    /** Root 블럭만 추출해 리스트로 변환 */
-    private static List<BlockResponseDto> getRootBlocks(List<Block> allBlocks, Map<Long, BlockResponseDto> dtoMap) {
-        return allBlocks.stream()
-                .filter(b -> b.getParent() == null)
-                .map(b -> dtoMap.get(b.getId()))
-                .toList();
-    }
-
-    /** BlockResponseDto 전용 맵 생성 */
-    private static Map<Long, BlockResponseDto> getBlockResponseDtoMap(List<Block> allBlocks) {
-        return allBlocks.stream()
-                .map(block -> BlockResponseDto.of(block, new ArrayList<>()))
-                .collect(Collectors.toMap(BlockResponseDto::getBlockId, dto -> dto));
-    }
-
-    /** 부모-자식 관계 연결 담당하는 전용 메서드 */
-    private void connectNodes(List<Block> allBlocks, Map<Long, BlockResponseDto> dtoMap) {
-        for (Block block : allBlocks) {
-            if (block.getParent() != null) { // 부모 블럭이 아니면
-                // 부모 블럭을 찾는다.
-                BlockResponseDto parentDto = dtoMap.get(block.getParent().getId());
-                if (parentDto != null) { // 부모 블럭이 있다면
-                    // 부모블럭에 현재 자식 블럭을 추가한다.
-                    parentDto.getChildren().add(dtoMap.get(block.getId()));
-                }
-            }
-        }
     }
 
     /** Tag 전용 맵 생성 */
