+++ src/main/java/com/joojoo/api/block/application/service/command/CreateBlockService.java
@@ -1,10 +1,10 @@
 package com.joojoo.api.block.application.service.command;
 
-import com.joojoo.api.block.application.detail.BlockDtoAssembler;
 import com.joojoo.api.block.application.port.in.CreateBlockUseCase;
 import com.joojoo.api.block.domain.model.entity.Block;
 import com.joojoo.api.block.domain.repository.BlockRepository;
 import com.joojoo.api.block.domain.service.BlockDomainService;
+import com.joojoo.api.block.domain.service.assembler.BlockTreeAssembler;
 import com.joojoo.api.block.domain.service.validation.BlockValidator;
 import com.joojoo.api.block.presentation.dto.request.createBlock.BlockSaveRequestDto;
 import com.joojoo.api.block.presentation.dto.response.blockDetail.BlockResponse;
@@ -26,9 +26,8 @@ public class CreateBlockService implements CreateBlockUseCase {
     private final GetUserUseCase getUserUseCase;
     private final BlockDomainService blockDomainService;
     private final BlockValidator blockValidator;
-
+    private final BlockTreeAssembler blockTreeAssembler;
     private final MetadataUseCase metadataUseCase;
-    private final BlockDtoAssembler blockDtoAssembler;
 
     @Override
     public BlockResponse saveBlockTree(Long userId, BlockSaveRequestDto requestDto) {
@@ -39,8 +38,7 @@ public class CreateBlockService implements CreateBlockUseCase {
         blockRepository.saveAll(allBlocks); // TODO : JDBC Batch Insert 고려, 지금 블럭이 10개면 10개의 Insert 쿼리 작동 중
 
         metadataUseCase.processMetadata(allBlocks, user.getId());
-        return blockDtoAssembler.assembleReconstructBlockTree(allBlocks);
+        return blockTreeAssembler.assembleReconstructBlockTree(allBlocks);
     }
-
 
 }
